requirements.txt
python-dotenv
openai  
pdfplumber